requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
docx2txt
openpyxl
